Work on integration test, add sql script

+++ back/mdd-core/src/main/java/com/mddcore/usecases/user/DeleteUserUseCase.java
@@ -26,12 +26,16 @@ public class DeleteUserUseCase extends UseCase<DeleteUserUseCase.InputValues, De
         User user = userRepository.findById(input.id()).orElseThrow(() ->
                 new IllegalArgumentException("User not found, cant delete it"));
 
+        if(!user.getId().equals(input.authId)) {
+            throw new IllegalArgumentException("User id don't match auth id");
+        }
+
         userRepository.delete(user);
 
         return new OutputValues(true);
     }
 
-    public record InputValues(Long id) implements UseCase.InputValues {
+    public record InputValues(Long id, Long authId) implements UseCase.InputValues {
     }
